Move to regular LogTree storage
The sequenced storage backend requires an index for all leafs added.

As we're indexing events from CloudTrail we could possibly derive one from the
timestamp but it seems easier to for now to use the regular backend and move
later if it makes sense.

In tests we are using the memory storage backend which makes it easy to run with
fewer dependencies.