Commit the transaction, don't Close it
The Close() function isn't actually part of the sql.Tx interface so it looks
like we haven't been cleaning ending transactions open to the database.

Adding in defer tx.Commit() resulted in some other problems (subsequent runs of
the lambda function would be unable to establish a new database connection) so
it seems best to call it inline with the function.